Improve class names.
