+++ module/VuFind/src/VuFind/Controller/AbstractBase.php
@@ -531,9 +531,9 @@ class AbstractBase extends AbstractActionController implements AccessPermissionI
      *
      * @param string $name Name of service to retrieve
      *
-     * @return \VuFind\Db\Service\AbstractService
+     * @return \VuFind\Db\Service\AbstractDbService
      */
-    public function getDbService(string $name): \VuFind\Db\Service\AbstractService
+    public function getDbService(string $name): \VuFind\Db\Service\AbstractDbService
     {
         return $this->serviceLocator->get(\VuFind\Db\Service\PluginManager::class)
             ->get($name);
