+++ module/VuFind/src/VuFind/Db/Row/User.php
@@ -70,10 +70,10 @@ class User extends RowGateway implements
     \VuFind\Db\Interface\UserAccountInterface,
     \VuFind\Db\Table\DbTableAwareInterface,
     \LmcRbacMvc\Identity\IdentityInterface,
-    \VuFind\Db\Service\ServiceAwareInterface
+    \VuFind\Db\Service\DbServiceAwareInterface
 {
     use \VuFind\Db\Table\DbTableAwareTrait;
-    use \VuFind\Db\Service\ServiceAwareTrait;
+    use \VuFind\Db\Service\DbServiceAwareTrait;
 
     /**
      * VuFind configuration
