+++ module/VuFind/src/VuFind/Db/Service/AbstractDbService.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Database service factory
+ * Database service abstract base class
  *
  * PHP version 8
  *
@@ -29,14 +29,8 @@
 
 namespace VuFind\Db\Service;
 
-use Interop\Container\ContainerInterface;
-use Interop\Container\Exception\ContainerException;
-use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
-use Laminas\ServiceManager\Exception\ServiceNotFoundException;
-use Laminas\ServiceManager\Factory\FactoryInterface;
-
 /**
- * Database service factory
+ * Database service abstract base class
  *
  * @category VuFind
  * @package  Database
@@ -44,27 +38,6 @@ use Laminas\ServiceManager\Factory\FactoryInterface;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:database_gateways Wiki
  */
-class AbstractServiceFactory implements FactoryInterface
+abstract class AbstractDbService
 {
-    /**
-     * Create an object
-     *
-     * @param ContainerInterface $container     Service manager
-     * @param string             $requestedName Service being created
-     * @param null|array         $options       Extra options (optional)
-     *
-     * @return object
-     *
-     * @throws ServiceNotFoundException if unable to resolve the service.
-     * @throws ServiceNotCreatedException if an exception is raised when
-     * creating a service.
-     * @throws ContainerException&\Throwable if any other error occurs
-     */
-    public function __invoke(
-        ContainerInterface $container,
-        $requestedName,
-        array $options = null
-    ) {
-        return new $requestedName(...($options ?? []));
-    }
 }
