+++ module/VuFind/src/VuFind/Db/Service/PluginManager.php
@@ -66,6 +66,6 @@ class PluginManager extends \VuFind\ServiceManager\AbstractPluginManager
      */
     protected function getExpectedInterface()
     {
-        return AbstractService::class;
+        return AbstractDbService::class;
     }
 }
