+++ module/VuFind/src/VuFind/Db/Service/UserServiceFactory.php
@@ -43,7 +43,7 @@ use Laminas\ServiceManager\Exception\ServiceNotFoundException;
  * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
  * @link     https://vufind.org/wiki/development:plugins:database_gateways Wiki
  */
-class UserServiceFactory extends AbstractServiceFactory
+class UserServiceFactory extends AbstractDbServiceFactory
 {
     /**
      * Create an object
